main.go: check IPv4 and IPv6 for domain from local machine and update
This commit covers IPv4 check, IPv6 check, error handling for each case,
and checking the existence of the domain within route53. If done properly,
exit code is zero. If there is an error (such as a domain not existing),
you will get a return code of 1. This could be slightly more optimized
to make a single call for running the update, but will not be explored
at this time. This gets basic funcationality up and running as well as
ensuring that we don't send/receive bad information along the way.

If you set your hostname in the os to be the FQDN, this program can
be run without any flags passed as long as you have set up your
AWS profile and/or exported your credentials. This follows standard
AWS SDK creds hierarchy, so if you have an instance profile, it will
pick that up as well.