Create function RetrieveResourceRecordSet to retrieve the resource record from route53 that needs modification
This is the initial function definition and execution. It does not currently perform all business logic
as it is an interim commit. Business logic will filter for exact resource record of specific type (A, AAAA, MX, CNAME, PTR, etc.) and return single record ID for modification.

TODO: filter resource records returned from call to just the resource record being requested and return
only the rrID of the FQDN, TYPE (A, AAAA), and value for comparison/update.

TODO: unit testing is currently done outside of git as IDs and resources are not yet mocked.
Once business logic is finished and unit tests are mocked. Unit tests will be merged into branch and mainline.